Refactored Embed Lexical card
